+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+ - ClientSerialRpcController - Adding hooks to access ConnectionStateManager states
 
 ## [7.5.2] - 2019-10-29
 ### Added